# docs/api-reference/openapi.yaml
openapi: 3.0.1
info:
    title: OpnForm API
    description: API for interacting with OpnForm, primarily used for Zapier integration
    version: 1.0.0
servers:
    - url: https://api.opnform.com
security:
    - bearerAuth: []
...
paths:
...
    "/open/workspaces/{workspaceId}/forms":
        get:
            tags:
                - Forms
            summary: List Workspace Forms
            description: "Retrieve all forms within a specific workspace. Returns lightweight form summaries (excludes `properties`). Use the Get Form endpoint to retrieve the full form with all fields. Requires `forms-read` ability."
            security:
                - bearerAuth: []
            parameters:
                - name: workspaceId
                  in: path
                  required: true
                  schema:
                      type: number
                      description: ID of the workspace to retrieve forms from.
                - name: page
                  in: query
                  required: false
                  schema:
                      type: number
                      minimum: 1
                      default: 1
                  description: "Pagination page number"
                - name: per_page
                  in: query
                  required: false
                  schema:
                      type: number
                      minimum: 1
                      maximum: 100
                      default: 10
                  description: "Number of forms per page"
            responses:
                "200":
                    description: Successful
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/FormListItem"
                                    links:
                                        type: object
                                        properties:
                                            first:
                                                type: string
                                            last:
                                                type: string
                                            prev:
                                                type: string
                                                nullable: true
                                            next:
                                                type: string
                                                nullable: true
                                    meta:
                                        type: object
                                        properties:
                                            current_page:
                                                type: number
                                            from:
                                                type: number
                                            last_page:
                                                type: number
                                            per_page:
                                                type: number
                                            to:
                                                type: number
                                            total:
                                                type: number
                "403":
                    description: "Forbidden – token lacks forms-read ability or you do not have access to the workspace"
                "404":
                    description: "Workspace not found"
    "/open/forms":
        get:
            tags:
                - Forms
            summary: List Forms
            description: "Retrieve all forms accessible to the token holder across all workspaces. Returns lightweight form summaries (excludes `properties`). **Deprecated**: Use `/open/workspaces/{workspaceId}/forms` instead."
            deprecated: true
            security:
                - bearerAuth: []
            responses:
                "200":
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/FormListItem"
                "403":
                    description: Forbidden – token lacks forms-read ability
        post:
            tags:
                - Forms
            summary: Create Form
            description: "Create a new form; requires forms-write ability."
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Form"
            responses:
                "201":
                    description: Form created
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Form"
                "403":
                    description: Forbidden – token lacks forms-write ability
    "/open/forms/{slug}":
        get:
            tags:
                - Forms
            summary: Get Form
            description: Retrieve a single form.
            security:
                - bearerAuth: []
            parameters:
                - name: slug
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Successful
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Form"
                "404":
                    description: Not Found
                "403":
                    description: Forbidden
    "/open/forms/{id}":
        put:
            tags:
                - Forms
            summary: Update Form
            description: "Update an existing form. Requires `forms-write`."
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: number
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Form"
            responses:
                "200":
                    description: Updated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Form"
                "403":
                    description: Forbidden – token lacks forms-write ability
        delete:
            tags:
                - Forms
            summary: Delete Form
            description: "Delete a form. Requires `forms-write`."
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: number
            responses:
                "204":
                    description: Deleted
                "403":
                    description: Forbidden
...
    "/external/zapier/forms":
        get:
            tags:
                - Zapier
            summary: List Forms
            description: Retrieve a list of forms available in a specific workspace.
            parameters:
                - name: workspace_id
                  in: query
                  description: The ID of the workspace for which to list forms
                  required: true
                  schema:
                      type: number
            responses:
                "200":
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Form"
...
components:
    schemas:
...
        Form:
            type: object
            properties:
                # Core form identification
                id:
                    type: number
                    description: The unique identifier for the form.
                    readOnly: true
                slug:
                    type: string
                    description: The URL-friendly slug for the form.
                    readOnly: true
                workspace_id:
                    type: number
                    description: ID of the workspace that owns the form.
                    writeOnly: true
                title:
                    type: string
                    description: The title of the form.
                visibility:
                    type: string
                    enum: [public, draft, closed]
                    description: The current visibility state of the form.
                tags:
                    type: array
                    items:
                        type: string
                    nullable: true
                language:
                    type: string
                    description: Two-letter ISO language code.
                custom_domain:
                    type: string
                    nullable: true

                # Visual appearance and styling
                theme:
                    type: string
                    enum: [default, simple, notion]
                font_family:
                    type: string
                    nullable: true
                color:
                    type: string
                dark_mode:
                    type: string
                    enum: [light, dark, auto]
                width:
                    type: string
                    enum: [centered, full]
                size:
                    type: string
                    enum: [sm, md, lg]
                border_radius:
                    type: string
                    enum: [none, small, full]
                layout_rtl:
                    type: boolean
                uppercase_labels:
                    type: boolean

                # Media and branding
                cover_picture:
                    type: string
                    format: uri
                    nullable: true
                logo_picture:
                    type: string
                    format: uri
                    nullable: true
                no_branding:
                    type: boolean
                    description: Whether to hide the OpnForm branding.
                transparent_background:
                    type: boolean
                    description: Transparent background when form is embedded.

                # Form behavior and submission settings
                submit_button_text:
                    type: string
                    maxLength: 50
                submitted_text:
                    type: string
                    maxLength: 2000
                redirect_url:
                    type: string
                    format: uri
                    nullable: true
                re_fillable:
                    type: boolean
                re_fill_button_text:
                    type: string
                    maxLength: 50
                confetti_on_submission:
                    type: boolean
                show_progress_bar:
                    type: boolean
                submission_retention_value:
                    type: integer
                    minimum: 1
                    maximum: 3650
                    nullable: true
                    description: Number of retention units before a submission is permanently deleted. Set this and submission_retention_unit to null to disable automatic deletion.
                submission_retention_unit:
                    type: string
                    enum: [day, week, month, year]
                    nullable: true
                    description: Calendar unit used by submission_retention_value. Both retention fields must be provided together.

                # Form closure and limits
                closes_at:
                    type: string
                    format: date-time
                    nullable: true
                closed_text:
                    type: string
                    nullable: true
                max_submissions_count:
                    type: integer
                    minimum: 1
                    nullable: true
                max_submissions_reached_text:
                    type: string
                    nullable: true

                # User experience features
                auto_save:
                    type: boolean
                auto_focus:
                    type: boolean
                enable_partial_submissions:
                    type: boolean
                partial_submission_abandonment_value:
                    type: integer
                    minimum: 1
                    maximum: 3650
                    nullable: true
                    description: Number of abandonment units after the last partial save before a draft is marked abandoned. Set this and partial_submission_abandonment_unit to null to keep drafts as partial indefinitely.
                partial_submission_abandonment_unit:
                    type: string
                    enum: [minute, hour, day]
                    nullable: true
                    description: Unit used by partial_submission_abandonment_value. Both fields must be provided together.
                editable_submissions:
                    type: boolean
                editable_submissions_button_text:
                    type: string
                    maxLength: 50

                # Security and validation
                password:
                    type: string
                    nullable: true
                use_captcha:
                    type: boolean
                captcha_provider:
                    type: string
                    enum: [recaptcha, hcaptcha]

                # SEO and indexing
                can_be_indexed:
                    type: boolean
                seo_meta:
                    type: object
                    nullable: true

                # Advanced features
                custom_code:
                    type: string
                    nullable: true
                database_fields_update:
                    type: array
                    items: {}
                    nullable: true
                properties:
                    type: array
                    description: An array of field and layout blocks that make up the form.
                    items:
                        $ref: "#/components/schemas/FormProperty"
        FormListItem:
            type: object
            description: A lightweight form summary returned by list endpoints. Use the Get Form endpoint to retrieve full form details including properties.
            properties:
                id:
                    type: number
                    description: The unique identifier for the form.
                    readOnly: true
                slug:
                    type: string
                    description: The URL-friendly slug for the form.
                    readOnly: true
                title:
                    type: string
                    description: The title of the form.
                visibility:
                    type: string
                    enum: [public, draft, closed]
                    description: The current visibility state of the form.
                tags:
                    type: array
                    items:
                        type: string
                    nullable: true
                views_count:
                    type: number
                    description: Total number of form views.
                submissions_count:
                    type: number
                    description: Total number of completed submissions.
                created_at:
                    type: string
                    format: date-time
                    description: ISO 8601 timestamp of when the form was created.
                updated_at:
                    type: string
                    format: date-time
                    description: ISO 8601 timestamp of when the form was last modified.
                last_edited_human:
                    type: string
                    description: Human-readable time since last edit (e.g., "2 hours ago").
                closes_at:
                    type: string
                    format: date-time
                    nullable: true
                    description: When the form will close for submissions.
                is_closed:
                    type: boolean
                    description: Whether the form is currently closed.
                max_submissions_count:
                    type: integer
                    minimum: 1
                    nullable: true
                    description: Maximum number of submissions allowed.
                max_number_of_submissions_reached:
                    type: boolean
                    description: Whether the submission limit has been reached.
                is_pro:
                    type: boolean
                    description: Whether the form's workspace has a Pro subscription.
                workspace_id:
                    type: number
                    description: ID of the workspace that owns the form.
                share_url:
                    type: string
                    format: uri
                    description: Public URL to share the form.
        FormProperty:
            type: object
            description: A single field or layout block within a form.
            properties:
                id:
                    type: string
                    description: A unique identifier for the block.
                    format: uuid
                type:
                    type: string
                    description: The type of block.
                name:
                    type: string
                    description: The label or name of the field.
                help:
                    type: string
                    description: "Help text displayed with the field."
                hidden:
                    type: boolean
                    description: Whether the field is hidden from view.
                required:
                    type: boolean
                    description: "Whether the field is required."
                placeholder:
                    type: string
                    description: Placeholder text for the input.
                width:
                    type: string
                    enum: [full, "1/2", "1/3", "2/3", "1/4", "3/4"]
                    description: The width of the block in the form layout.
            additionalProperties: true
...
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: "Personal Access Token"
            x-bearer-scopemap:
                "workspaces-read": "Read access to workspaces"
                "workspaces-write": "Write access to workspaces"
                "workspace-users-read": "Read access to workspace users"
                "workspace-users-write": "Write access to workspace users"
                "forms-read": "Read access to forms"
                "forms-write": "Write access to forms"
                "manage-integrations": "Manage form integrations (webhooks)"